Renamed Cursor to Cursors to there's not Namespace ambiguity when using UnityEngine.Cursor. Removed scene from build settings.
