+++ Assets/MixedRealityToolkit/UX/Scripts/Cursors/AnimatedCursor.cs
@@ -4,7 +4,7 @@
 using System;
 using UnityEngine;
 
-namespace MixedRealityToolkit.UX.Cursor
+namespace MixedRealityToolkit.UX.Cursors
 {
     /// <summary>
     /// Animated cursor is a cursor driven using an animator to inject state information
