+++ Assets/MixedRealityToolkit/UX/Scripts/Cursors/InteractiveMeshCursor.cs
@@ -4,7 +4,7 @@
 using MixedRealityToolkit.InputModule.Gaze;
 using UnityEngine;
 
-namespace MixedRealityToolkit.UX.Cursor
+namespace MixedRealityToolkit.UX.Cursors
 {
     /// <summary>
     /// A cursor that looks and acts more like the shell cursor.
