+++ Assets/MixedRealityToolkit/UX/Scripts/Cursors/MeshCursor.cs
@@ -4,7 +4,7 @@
 using System;
 using UnityEngine;
 
-namespace MixedRealityToolkit.UX.Cursor
+namespace MixedRealityToolkit.UX.Cursors
 {
     /// <summary>
     /// Object that represents a cursor in 3D space controlled by gaze.
