+++ ProjectSettings/EditorBuildSettings.asset
@@ -4,7 +4,4 @@
 EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
-  m_Scenes:
-  - enabled: 1
-    path: Assets/MixedRealityToolkit-Examples/Input/Scenes/MicrophoneStream.unity
-    guid: 0d1edddc695c93840945d97459c6f5cc
+  m_Scenes: []
